fix machine status updater
